+++ lib/redmine_monitoring/notifications/channels/email.rb
@@ -7,6 +7,11 @@ module RedmineMonitoring
         def self.deliver(payload)
           settings = Setting.plugin_redmine_monitoring || {}
           list = (settings['notify_email_recipients'] || '').split(/[,\s]+/).reject(&:blank?)
+          if list.blank?
+            RedmineMonitoring::OperationalLogger.once(:email_recipients_missing,
+                                                      level: :warn,
+                                                      message: 'email channel skipped: recipients missing')
+          end
           return if list.blank?
 
           MonitoringNotifierMailer.alert(payload, list).deliver_later